SocketIO and Relay server added
As from inside the browser we can only run a client and not a server, we have to introduce am intermediaire relay server.
The web app connects to this server and send the face pose data. Then the server forwards it to any connected client, such as UE.

Relay server must be running before launching any client. There is a batch for it...